+++ tdmclient/__init__.py
@@ -1,5 +1,5 @@
 # This file is part of tdmclient.
-# Copyright 2021 ECOLE POLYTECHNIQUE FEDERALE DE LAUSANNE,
+# Copyright 2021-2022 ECOLE POLYTECHNIQUE FEDERALE DE LAUSANNE,
 # Miniature Mobile Robots group, Switzerland
 # Author: Yves Piguet
 #
@@ -12,6 +12,7 @@ Communication with TDM.
 from tdmclient.fb import FlatBuffer, Table, Union
 from tdmclient.zeroconf import TDMZeroconfBrowser
 from tdmclient.tcp import TDMConnection
+from tdmclient.ws import TDMConnectionWS
 from tdmclient.thymio import ThymioFB, Node
 from tdmclient.client import Client
 from tdmclient.clientasync import ClientAsync