+++ tdmclient/tcp.py
@@ -1,5 +1,5 @@
 # This file is part of tdmclient.
-# Copyright 2021 ECOLE POLYTECHNIQUE FEDERALE DE LAUSANNE,
+# Copyright 2021-2022 ECOLE POLYTECHNIQUE FEDERALE DE LAUSANNE,
 # Miniature Mobile Robots group, Switzerland
 # Author: Yves Piguet
 #
@@ -82,11 +82,8 @@ class InputThread(threading.Thread):
 
 
 class TDMConnection:
-    """Connection to TDM.
+    """Connection to TDM via plain TCP.
     """
-
-    class TDMConnectionError(OSError):
-        pass
 
     def __init__(self,
                  host: str, port: int,