+++ tdmclient/tools/gui.py
@@ -19,7 +19,8 @@ class VariableTableWindow(tk.Tk):
 
     def __init__(self,
                  zeroconf=None,
-                 tdm_addr=None, tdm_port=None, password=None,
+                 tdm_addr=None, tdm_port=None, tdm_ws=False,
+                 password=None,
                  node_id=None, node_name=None,
                  language=None, debug=0):
         super(VariableTableWindow, self).__init__()
@@ -31,6 +32,7 @@ class VariableTableWindow(tk.Tk):
         self.zeroconf = zeroconf
         self.tdm_addr = tdm_addr
         self.tdm_port = tdm_port
+        self.tdm_ws = tdm_ws
         self.password = password
         self.node_id = node_id
         self.node_name = node_name
@@ -195,7 +197,7 @@ class VariableTableWindow(tk.Tk):
     def set_title(self):
         name = self.node.props["name"] if self.node is not None else "No robot"
         if self.client is not None and self.client.tdm_addr is not None:
-            name += f" (TDM: {self.client.tdm_addr}:{self.client.tdm_port})"
+            name += f" (TDM: {'ws://' if self.tdm_ws else ''}{self.client.tdm_addr}:{self.client.tdm_port})"
         if self.text_program is not None:
             name += " - "
             name += os.path.basename(self.program_path) if self.program_path is not None else f"Untitled.{self.language}"
@@ -482,6 +484,7 @@ class VariableTableWindow(tk.Tk):
         self.client = ClientAsync(zeroconf=self.zeroconf,
                                   tdm_addr=self.tdm_addr,
                                   tdm_port=self.tdm_port,
+                                  tdm_ws=self.tdm_ws,
                                   password=self.password,
                                   debug=self.debug)
         self.client.on_nodes_changed = on_nodes_changed
@@ -513,6 +516,7 @@ Options:
   --robotname=N  robot name; default=any
   --tdmaddr=H    tdm address (default: localhost or from zeroconf)
   --tdmport=P    tdm port or "default" for {ClientAsync.DEFAULT_TDM_PORT} (default: from zeroconf)
+  --tdmws        connect to tdm with WebSocket (default: plain TCP)
   --zeroconf     use zeroconf (default: automatic)
 """)
 
@@ -524,6 +528,7 @@ if __name__ == "__main__":
     zeroconf = None  # auto
     tdm_addr = None
     tdm_port = None
+    tdm_ws = False
     password = None
     robot_id = None
     robot_name = None
@@ -541,6 +546,7 @@ if __name__ == "__main__":
                                               "robotname=",
                                               "tdmaddr=",
                                               "tdmport=",
+                                              "tdmws",
                                               "zeroconf",
                                           ])
     except getopt.error as err:
@@ -566,11 +572,13 @@ if __name__ == "__main__":
             tdm_addr = val
         elif arg == "--tdmport":
             tdm_port = ClientAsync.DEFAULT_TDM_PORT if val == "default" else int(val)
+        elif arg == "--tdmws":
+            tdm_ws = True
         elif arg == "--zeroconf":
             zeroconf = True
 
     win = VariableTableWindow(zeroconf=zeroconf,
-                              tdm_addr=tdm_addr, tdm_port=tdm_port,
+                              tdm_addr=tdm_addr, tdm_port=tdm_port, tdm_ws=tdm_ws,
                               password=password,
                               node_id=robot_id, node_name=robot_name,
                               language=language, debug=debug)