+++ tdmclient/tools/run.py
@@ -37,6 +37,7 @@ Options:
   --stop         stop program (no filename or stdin expected)
   --tdmaddr=H    tdm address (default: localhost or from zeroconf)
   --tdmport=P    tdm port or "default" for {ClientAsync.DEFAULT_TDM_PORT} (default: from zeroconf)
+  --tdmws        connect to tdm with WebSocket (default: plain TCP)
 """, **kwargs)
 
 
@@ -48,6 +49,7 @@ if __name__ == "__main__":
     scratchpad = 0  # 1=--scratchpad, 2=--sponly
     tdm_addr = None
     tdm_port = None
+    tdm_ws = False
     password = None
     robot_id = None
     robot_name = None
@@ -98,6 +100,7 @@ if __name__ == "__main__":
                                               "stop",
                                               "tdmaddr=",
                                               "tdmport=",
+                                              "tdmws",
                                           ])
     except getopt.error as err:
         print(str(err), file=sys.stderr)
@@ -141,6 +144,8 @@ if __name__ == "__main__":
             tdm_addr = val
         elif arg == "--tdmport":
             tdm_port = ClientAsync.DEFAULT_TDM_PORT if val == "default" else int(val)
+        elif arg == "--tdmws":
+            tdm_ws = True
 
     if stop:
         if len(values) > 0:
@@ -199,7 +204,7 @@ if __name__ == "__main__":
         for event_name in transpiler.events_out:
             events.append((event_name, transpiler.events_out[event_name]))
 
-    with ClientAsync(tdm_addr=tdm_addr, tdm_port=tdm_port,
+    with ClientAsync(tdm_addr=tdm_addr, tdm_port=tdm_port, tdm_ws=tdm_ws,
                      password=password,
                      debug=debug) as client: