+++ tdmclient/tools/watch.py
@@ -22,6 +22,7 @@ Options:
   --robotname=N  robot name; default=any
   --tdmaddr=H    tdm address (default: localhost or from zeroconf)
   --tdmport=P    tdm port or "default" for {ClientAsync.DEFAULT_TDM_PORT} (default: from zeroconf)
+  --tdmws        connect to tdm with WebSocket (default: plain TCP)
 """, **kwargs)
 
 
@@ -30,6 +31,7 @@ if __name__ == "__main__":
     debug = 2  # display all messages received from the tdm by default
     tdm_addr = None
     tdm_port = None
+    tdm_ws = False
     password = None
     robot_id = None
     robot_name = None
@@ -45,6 +47,7 @@ if __name__ == "__main__":
                                               "robotname=",
                                               "tdmaddr=",
                                               "tdmport=",
+                                              "tdmws",
                                           ])
     except getopt.error as err:
         print(str(err), file=sys.stderr)
@@ -65,12 +68,14 @@ if __name__ == "__main__":
             tdm_addr = val
         elif arg == "--tdmport":
             tdm_port = ClientAsync.DEFAULT_TDM_PORT if val == "default" else int(val)
+        elif arg == "--tdmws":
+            tdm_ws = True
 
     if len(values) > 0:
         help(file=sys.stderr)
         sys.exit(1)
 
-    with ClientAsync(tdm_addr=tdm_addr, tdm_port=tdm_port,
+    with ClientAsync(tdm_addr=tdm_addr, tdm_port=tdm_port, tdm_ws=tdm_ws,
                      password=password,
                      debug=debug) as client: